pyproject.toml
[project]
name = "ANL488_FYP"
version = "0.1.0"
description = "ML project for Factory price prediction"
requires-python = ">=3.12"
dependencies = []

[project.optional-dependencies]
backend = [
    "pandas==2.2.3",
    "numpy==2.4.3",
    "scikit-learn==1.8.0",
    "joblib==1.5.3",
    "scipy==1.15.2",
    "xgboost==2.1.4",
    "fastapi==0.115.14",
    "uvicorn==0.34.3",
    "hydra-core==1.3.2",
    "pydantic-settings==2.9.1",
]
frontend = [
    "streamlit==1.56.0",
    "requests==2.32.3",
]
dev = [
    "matplotlib==3.10.8",
    "seaborn==0.13.2",
    "pyyaml==6.0.3",
    "statsmodels==0.14.4",
    "jupyter==1.1.1",
    "nbconvert==7.16.6",
    "ruff==0.11.2",
    "pytest==8.3.5",
    "mlflow==3.11.1",
    "hydra-optuna-sweeper==1.2.0",
    "optuna==2.10.1",
]
all = [
    "ANL488_FYP[backend]",
    "ANL488_FYP[frontend]",
    "ANL488_FYP[dev]",
]

[tool.ruff]
line-length = 120
exclude = ["notebooks/", ".venv_BMAD/", ".claude/", "data/", "models/", "others/"]

[tool.ruff.lint]
select = ["E", "F", "W", "I"]
